refactor, use friendly-errors-webpack-plugin, remove watch mode
