+++ lib/get-mode.js
@@ -4,9 +4,6 @@ module.exports = function (options) {
   if (options.electron) {
     modes.push('Electron')
   }
-  if (options.watch) {
-    modes.push('Watch')
-  }
   if (options.dev) {
     modes.push(options.live ? 'Live Reloading' : 'Hot Reloading')
   }
