+++ lib/hmr.js
@@ -12,13 +12,13 @@ const addEntry = (entry, newEntry) => {
 
 module.exports = function (config, options) {
   if (options.dev) {
-    const publicPath = `http://${options.host === 'all' ? 'localhost' : options.host}:${options.port}/`
+    const publicPath = `http://${options.host}:${options.port}/`
     // with electron and dev mode
-    if (!options.watch && options.electron) {
+    if (options.electron) {
       config.output.publicPath = publicPath
     }
     // with dev mode and hot reloading
-    if (!options.watch && !options.live) {
+    if (!options.live) {
       const pathToHotMiddleware = require.resolve('webpack-hot-middleware/client') + '?reload=true'
 
       let hmrClient
