+++ lib/index.js
@@ -6,7 +6,6 @@ const chalk = require('chalk')
 const merge = require('lodash.merge')
 const tildify = require('tildify')
 const updateConfig = require('./update-config')
-const terminalUtils = require('./terminal-utils')
 const defaultBabelOptions = require('./webpack/babel-options')
 const requireWithContext = require('./require-with-context')
 
@@ -37,9 +36,6 @@ module.exports = co.wrap(function * (options) {
     }
   }
 
-  if (options.watch) {
-    options.dev = true
-  }
   const isDev = options.dev
 
   process.env.NODE_ENV = isDev ? 'development' : 'production'
@@ -56,7 +52,7 @@ module.exports = co.wrap(function * (options) {
     {
       port,
       notify: true,
-      host: 'all',
+      host: 'localhost',
       dist: 'dist',
       babel: defaultBabelOptions,
       static: {
@@ -81,11 +77,10 @@ module.exports = co.wrap(function * (options) {
     webpackConfig = require('./merge-config')(webpackConfig, buildOptions)
   }
 
-  terminalUtils.clearConsole()
   if (buildOptions.test) {
     require('./test')(webpackConfig, buildOptions)
-  } else if (buildOptions.dev && !buildOptions.watch) {
-    yield require('./dev')(webpackConfig, buildOptions)
+  } else if (buildOptions.dev) {
+    require('./dev')(webpackConfig, buildOptions)
   } else {
     yield require('./build')(webpackConfig, buildOptions)
   }
