+++ lib/server.js
@@ -5,6 +5,7 @@ const webpack = require('webpack')
 const httpProxyMiddleware = require('http-proxy-middleware')
 const historyApiFallback = require('connect-history-api-fallback')
 const chalk = require('chalk')
+const opn = require('opn')
 
 module.exports = function (webpackConfig, options) {
   const app = express()
@@ -37,6 +38,20 @@ module.exports = function (webpackConfig, options) {
     hpm = addProxyMiddleware(app, proxy)
   }
 
+  const port = options.port
+  const host = options.host
+  const url = `http://${host}:${port}`
+
+  devMiddleWare.waitUntilValid(() => {
+    console.log(`> Running on ${url}`)
+    if (host === '0.0.0.0') {
+      console.log(chalk.dim(`To visit it on other device, connect your devices to the same network and open http://${require('internal-ip').v4()}:${port}`))
+    }
+    if (options.open) {
+      opn(url)
+    }
+  })
+
   app.get('*', (req, res) => {
     devMiddleWare.waitUntilValid(() => {
       const html = mfs.readFileSync(file)
@@ -44,9 +59,7 @@ module.exports = function (webpackConfig, options) {
     })
   })
 
-  const port = options.port
-  const host = options.host
-  const appServer = app.listen(port, host === 'all' ? '0.0.0.0' : host)
+  const appServer = app.listen(port, host)
 
   if (proxy) {
     // Listen for the websocket 'upgrade' event and upgrade the connection.
@@ -91,8 +104,8 @@ function onProxyError(proxy) {
 }
 
 function addProxyMiddleware(devServer, proxy) {
-  //if proxy is specified we fallback to index.html if request `accept`s text/html
-  //so proxy works with vue-router or other client routing libraries
+  // if proxy is specified we fallback to index.html if request `accept`s text/html
+  // so proxy works with vue-router or other client routing libraries
   devServer.use(historyApiFallback({
     disableDotRule: true,
     htmlAcceptHeaders: ['text/html']
