+++ lib/webpack/config.dev.js
@@ -1,5 +1,6 @@
 'use strict'
 const webpack = require('webpack')
+const FriendlyErrors = require('friendly-errors-webpack-plugin')
 const config = require('./config')
 
 config.devtool = 'eval-source-map'
@@ -9,8 +10,9 @@ config.plugins = config.plugins.concat([
   new webpack.DefinePlugin({
     __DEV__: true,
     'process.env.NODE_ENV': JSON.stringify('development')
-  })
+  }),
   /*eslint-enable */
+  new FriendlyErrors()
 ])
 
 module.exports = config
