+++ lib/webpack/config.js
@@ -1,6 +1,5 @@
 'use strict'
 const _ = require('../utils')
-const DisplayStats = require('./plugins/display-stats')
 
 module.exports = {
   context: _.cwd(),
@@ -49,7 +48,5 @@ module.exports = {
       // css loader will be added in ./update-config.js
     ]
   },
-  plugins: [
-    new DisplayStats()
-  ]
+  plugins: []
 }
