+++ lib/webpack/get-stats.js
@@ -23,7 +23,6 @@ module.exports = stats => {
     chunks: false,
     chunkModules: false,
     hash: false,
-    version: false,
     timings: false
   }) + successMessage
 }
