+++ package.json
@@ -76,6 +76,7 @@
     "express": "^4.14.0",
     "extract-text-webpack-plugin": "^2.0.0-beta.3",
     "file-loader": "^0.9.0",
+    "friendly-errors-webpack-plugin": "^1.1.2",
     "fs-promise": "^0.5.0",
     "graceful-copy": "^1.0.1",
     "html-webpack-plugin": "^2.22.0",
@@ -94,7 +95,6 @@
     "rimraf": "^2.5.4",
     "source-map-support": "^0.4.6",
     "style-loader": "^0.13.1",
-    "text-table": "^0.2.0",
     "tildify": "^1.2.0",
     "time-stamp": "^1.0.1",
     "update-notifier": "^1.0.2",
